app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
flasgger==0.9.7.1
Flask==3.0.3
flask-pydantic-api==1.3.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
